Revert "Revert "only force off KEYSTONE_USE_WSGI up until kilo""
This reverts commit ccbfaa83ac82f14a9d7d9cc658d2899d41fc9dda.

This was never the correct fix, it just masked the root problem. This
correctly was setting the variables but there was a bug in grenade that
made this not do the right thing on stable/kilo because of a missing 
source include.

That's where it should have been fixed.

Related-Bug: #1456835

Change-Id: I557e9d7d35239b9336b6d90f037f14cdf2585f76