fix grenade phase check
The check for which phase grenade was working on was based on checking
the contents of a variable passed in to setup_localrc file. However,
this was changes from localrc_old to localrc_oldnew sometime during
the multinode work, but one call to it remained. That meant all
devstacks were told they were GRENADE_PHASE=target.

It didn't seem to be hurting anything, but we should fix this fact.

Change-Id: I1a4b888a493c0b99200b2574d279039e3d5f4ef3