Add option to setup netconsole to remote destination
Debugging a kernel oops or other issues where jenkins drops connection
with the host is very difficult.  The console screen requires one of
the friendly infra-* staff to examine, but even then it's small and
can scroll off important information or blank out.  Also, if you're
killing the network connection you're left with no information what
might be going on in the host.

After hitting this issue recently, I did some change-specific stuff to
setup a netconsole on the host and send the logs out that way.  It
ended up working well and gave me a full backtrace of the oops, which
cut down the debugging time considerably.

This wraps the setup into a function so you just need to specify
DEVSTACK_GATE_NETCONSOLE=123.123.123.123:6666 and it will setup the
host to send logs there.  It is intended for short-term use while
you're debugging specific problems with a job.

Change-Id: Iebc8e1812d1441aba7c18d3e7c982b620b5198a0