Fix the setup_workspace to handle missing ZUUL_BRANCH
This commit adds logic to the setup_workspace function to be able to
handle the case where the var ZUUL_BRANCH isn't set. This is what
apparently happens during periodic jobs and causes those to fail.
If ZUUL_BRANCH isn't set we just use the fallback zuul ref which is
what the expected behavior would be if the function didn't explode.

Change-Id: I77b98b8a8d8602da0092a5250cc0ce49de9215db