Correct case pattern handling in MOD_WSGI checks
If we've got two identical patterns one of them needs to fall
through or have them both in the same block.

It appears there was some merge confusion with
f42d6c8a8d39cfc65fb1f5804f84608a0e7eab59 and
5f8b3cb631688ec65629fffaf0f8b7615a16c511

This tries to get the correct the original intent: flat out disable
for ceilometer and disable depending on branch for keystone.

Change-Id: I9e5e792ef43f9a3cdc0b84c73f27b62109ed5d31