Remove q-vpn service for DSVM tests
There is a DevStack plugin for the neutron-vpnaas repo now, that
enables the q-vpn service. The repo has DSVM based (functional)
tests that use the new plugin and exercise the service.

We now want to remove q-vpn setup from the DevStack repo, and to
rely on the new plugin. To do that, however, we need to make sure
that DSVM tests (other than those in the neutron-vpnaas repo) do
not enable q-vpn, as it will not be set up during stacking (and
q-l3 will not be properly set up if q-vpn is enabled).

This patch, along with I3a4d7ecb25668eb499677d05bdb68b919592e53d
are attempting to remove q-vpn from the list of enabled services.

Change-Id: Idd5524ed0b602408be5a53830981a8ab974b390c