Revert "create partition for lvm volume"
This reverts commit 034189ec041aca5740c1689f3034f9b2e9d4418e.

This change was made with the justification that it would make
cinder use "a bit more optimized" and that it "may impact reliability
for good".  There was a link in a review comment to a bug that did
not appear to have any relevance.

In short, this change added both complexity and essentially a test
environment requirement for a second volume.  Perhaps we need that,
but if we are going to establish that as a requirement, we should
have some evidence to back up the performance claims.

Change-Id: I541a66d81fab5be65bb9cd6684e00aa85c3283a9