grenade-multinode: don't set upgrade level initially
In Nova, services talk the latest version of RPC, unless you tell them
otherwise. This can be set with the upgrade_level config option.

You are supposed to set this *before* upgrade, not *before* initial
installation. While this happened to work in the kilo case, it was the
wrong place to do it. Grenade itself will be responsible for setting
this at the correct time in I5c899779866ebdf492d474a625eb396273bb0259
and devstack-gate should not be in the business of touching this.

Change-Id: Ieb92f52180369f2513bda40dcbaa9db392816841