set min_compute_nodes when running in multinode mode
To ensure that our multinode configurations actually have multiple
working workers, we're building some tests in Tempest that attempt to
land on every node by using scheduler hints (force_hosts). This
requires that we set the minimum number of nodes that are expected to
be found in a configuration (it defaults to 1).

Change-Id: I59ca8b76dc3fe0c62236de11dbf44d52fa37226a
Depends-On: I7d7230d100901179a1d1a0281aa8be990d00c02a