Fix grenade on OVH with nova net
We had previously forced nova net with just devstack to not touch eth0
as it is unable to reconfigure it properly on OVH. Instead we use an
internal bridge which is detached from the test host's networking.
However we wrote the config to the wrong localrc when running grenade on
OVH.

Fix this by properly selecting the localrc file based on the test mode
variable. This should write the new networking config to the proper
config file separating grenade + nova-net's network from the test host's
networking.

Change-Id: I7d10f45a30efd1a9d6587cc5e089a7340b144311