Bump ironic test timeout to 900
We're still failing occassionally; data tells us that when we get a slow
host, it's failing during boot on the last cleaning phase.

We have found that it's mostly userspace taking so long at boot, so
we're going down the path of making that faster.

Bump the timeout while we work on that.

Change-Id: I9287811d7cdc10e00b3e59c4e6adef8bbe233d2b
Related-Bug: #1393099