Install and run StackViz during devstack-gate host cleanup
StackViz is a utility for visualizing the performance of DevStack and
Tempest executions. This adds a new function, `process_stackviz`, that
copies a static site template built during nodepool image build (see
dependent patch), and then installs and runs a small Python module
(already cached on the filesystem) to process subunit data. This data,
along with the static template, is then copied into the log output
directory so that it can be uploaded to the log server when the job
finishes.

Change-Id: Ia5bee69f2874ca50ac6148494d2926ef4ab73b6b
Depends-On: I9bc50b8eba696264d2b97ed20a37f28f4768abb8