Use journald when it is present
Previously we only assumed journald was present on fedora, but now it
exists on centos7 and ubuntu xenial. Instead of using the distro to
determine if journald is present test for the presence of journalctl.

If we don't hvae a journalctl fall back to the old behavior and assume
rsyslog is in use.

Change-Id: I4e1a3bdda7d46722b0a16ebb0b442917d4cb9b78