Print installed tempest-plugins after tempest run
The ALL_PLUGINS tox job runs tempest in a venv with system site-packages
enabled. This can make it very confusing to figure out exactly which
plugins are installed when tempest is run. Especially given many
projects' terrible choice to bundle the plugin in the project repo.
This commit adds running 'tempest list-plugins' inside the venv tox
makes for that job to show at the end of a run all the plugins tempest
was run with.

Change-Id: I8a65c851cdf5a4081ac32afa17f69ee001ff40c7