Enabling Cinder multinode testing
Cinder team is planning to start multinode testing in Newton cycle. We
want to be able to test volume migration/retype and partial grenade.

This commit enables c-vol and c-bak on the subnode in regular multinode
setups (this should work as I4d2960d92f388ac689dfa6b436dc8bfc1e129fbf is
merged now). Also a new features.yaml elements are introduced that would
disable c-vol and c-bak on primary node. This will be required for
multinode grenade jobs that will test compatibility of master c-api and
c-sch with stable c-vol and c-bak.

Co-Authored-By: Karthik Prabhu Vinod <karthik.prabhu.vinod@intel.com>

Change-Id: I4532bdcfa845cf0a405f68371c61559b0a70f9c1