Remove dib-utils & diskimage-builder from PROJECTS
No projects using devstack-gate currently install dib from git; these
are actually more like the python-* clients in that they are just
installed via pip as dependencies.

I do have a change out to the nodepool dsvm job to use dib from git
(480dcb201aa93bc3d208e2bf5e96bcee6dd7df09) but since that is the only
place we are doing this, we can add it to PROJECTS in that job and
clean it up from here.

Change-Id: Ia2b39193f481c879b454d10610ca1135959241b1