Put the regex first
This commit swaps the argument order of tempest tox jobs that use a
regex to specify a subset of tests to run. This is necessary for some
future changes to the tempest tox definitions that will switch it from
using a testr wrapping bash script to the tempest run command instead.
With tempest run you specify a regex with a '--regex' parameter and to
keep backwards compatibility this is added to the tempest tox
definitions. However this breaks on devstack-gate because we pass the
concurrency in before the regex. This ends up having a call equivalent
to:

  tempest run --regex --concurrency=2 $REGEX

which obviously is incorrect. Simply switching the arg order should work
in both cases. It's also worth pointing out that this is a temporary
step as we will migrate devstack-gate to call tempest init and tempest
run directly.

Change-Id: I8786b9709584378aaf48393115dda94397e22544
Need-by: I3684fce66a799579fa68af119652cafef25a9f03