Introduce roles into the feature matrix
test-matrix.py has a new cli option, for which role the node will play.
If it is a subnode, the starting set of services will be emitted,
that matches the current logic.

Logic regarding flags like DEVSTACK_GATE_NEUTRON,
DEVSTACK_GATE_NEUTRON_DVR, DEVSTACK_GATE_IRONIC is now moved to the
feature matrix.

Co-Authored-By: Julia Varlamova <jvarlamova@mirantis.com>
Co-Authored-By: Michal Dulko <michal.dulko@intel.com>

Change-Id: Ief1b74cbc01235d9f52d08add930781f3a0ae71b