Add GRENADE_PLUGINRC and TEMPEST_CONCURRENCY to reproduce.sh
For jobs that are running a grenade plugin (for example the Ironic
grenade job), the environment variable GRENADE_PLUGINRC is used.
Therefore if that variable exists it needs to be part of reproduce.sh.

If TEMPEST_CONCURRENCY has been explicitly set to 1, as is the case
for Ironic grenade jobs, then also write it to reproduce.sh.

Change-Id: Ib038053633e8b43430c8116daa27b1d6f5bce913