Add ansible callback plugin to make output nicer
Based on the in-tree minimal plugin, this reduces the amount of output,
and also adds timestamps. It also automatically outputs stdout, so we
can skip adding that ourselves. Finally - it gives us the framework to
poke with it, should we desire to morph things further.

The assumption is that CWD is $WORKSPACE when we're executing
ansible-playbook

Change-Id: I8378880d1a722be21b61311f3979483909b87fea