Remove oslo projects from PROJECTS list
By default we consume oslo libs from pypi and do not install them from
source repos. Because of this they shouldn't be in the default PROJECTS
list which is for configuring git repos to perform source installs from
them.

The handling for LIBS_FROM_GIT should mean that the oslo projects are
added as necessary to existing jobs that do install oslo from source.

Change-Id: I55c7d0e9a3f4be5c98e2ef19b38cd56416ca32ab