Fix tempest all jobs
The all jobs on tempest are supposed to run all tests without a regex
filter. However this was broken by (I really wish I could say
recently) tempest change I3684fce66a799579fa68af119652cafef25a9f03
which switched to tempest run. Since the primary use of the all tox
target there is plugin jobs which almost always run with a regex we
added a '--regex' argument there for backwards compatibility to not
break other projects. This however broke the all job because a --regex
needs an argument. This commit fixes this by adding the "regex" tempest
which will match all tests in the tempest repo and have the same net
effect as not specifying a selection regex at all.

Change-Id: I855a8ec30ed45f0b8227325f76bf48bd9e1e5988