Redirect stderr to stdout for some ansible calls
As described in the new in-line comment, the point of
devstack-early.txt is to capture output from stack.sh before it
switches itself to devstack.txt.

When stack.sh starts, it just does "set +x" ... this means tracing is
coming out on stderr.  ansible consumes all out and then, when the
command finishes, prints stdout and *then* stderr.  This is really
quite useless for tracing purposes, where obviously you want the
tracing interspaced with any command output.

To avoid this, redirect stderr to stdout on these ansible calls.  This
will ensure everything comes out sequentially.

Change-Id: Ieca949407bb371f0ad58ca8d4c9fa692b68e6e87