Specify etcd download location for all CI jobs
In Ibbb430fb1dbf66942168e0cb52d990ab6a2eb8d7 change for devstack,
we download etcd from github for end user scenarios. However this
is not acceptable for our CI. So we need to set the environment
variable for devstack to something in our CI infrastructure. We
need to do this such that it is available to all jobs as etcd
will run for all jobs from now on.

NOTE: We can get rid of the wget in devstack when the distro(s)
package etcd 3.x for general use.

Change-Id: If06731ed540072e78e7fceacf4ebf650ed89f3ea