Enable process_stackviz function to be optional
This patch keeps the proccess_stackviz function default to true, but
allow it to be optional called, if you set PROCESS_STACKVIZ to 0.
This is required because not all the projects run tempest in the same
way.
One of the examples is the tripleo project, that has a particular way
to run tempest (in undercloud) and it's being impossible to have the
stackviz results in the logs/stackviz because the cleanup process
overwrite the stackviz data there.

Change-Id: Idf4f360ecba618900ef939d94db2ffb73e16b11a