Group all horizon-related projects together
If django_openstack_auth or manila-ui are the projects being tested,
DEVSTACK_GATE_HORIZON may not be explicitly set for some gate jobs. For
instance, the tempest-dsvm-neutron-src job is run for
django_openstack_auth. Special case horizon-related projects to make
sure horizon is set up for them. Also further condense these projects in
the PROJECTS list to reduce the number of git repos we need to configure
in the default case

Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>
Change-Id: I384e4c376076051e7a4a8b2d704107cbe3ac08ad
Depends-On: Iafebfc92017d859cd3062fc4b6fab2986dfc5c98