Make sure we always run the nova metadata service
In the past we had gate configs where we didn't run the metadata service
because it wasn't as reliable as a config drive and it just wasted
memory. However, that isn't the case today, and almost all the jobs use
the metadata service. The exception here being ironic jobs, but running
the metadata service won't break things, it just will go unused. To
simplify the config and make the metadata api service config consistent
with the other nova services this commit makes sure that we are always
enabling the service in devstack.

Change-Id: I8ff08d61520ccf04e32dcd02f4cecc39dae823cb