Use caching reverse proxy for pypi
Test whether or not the pypi reverse proxy cache functions as expected.
We will want to keep an eye on two datapoints in particular. Reliablity
of devstack-gate with the reverse proxy cache rather than afs cache and
job runtimes.

We put this in d-g rather than on the images directly to have an easy
revert path. If we decide to do this long term because it is faster
and/or more reliable we will want to edit out images instead.

Change-Id: I92b63324b1f1c7bd5a41063f56c982ef7c1f1b01