Remove IRONIC_KEY_FILE
According to codesearch, this variable is not used anywhere.  The
key file is likely to be used by the generic switch driver, so
keep that variable but point it at the standard id_rsa key.  The
ironic_key is merely another copy of it; it's not clear why it
was a different file to start with.

Change-Id: If67c3fa1d7a5823ee0b0fe29c6902d08f336a299