Re-introduce sync/partprobe in fix_disk_layout
We've seen at least one occurence of a
"Device or resource busy while setting up superblock" error while
setting up xvde2 which, as far as we know, is a new error.

This problem was not happening before we took out sync/partprobe so
let's put that back in for safe measures.

Related-bug: #1706103
Change-Id: Iad1a75aa5fb4ae5a2f4a9cad6373091124917c24