Check for branches in zuulv3 git repos properly
Zuulv3 git repos do not have remotes. git_has_branch checks if the
branch exists in the origin remote which does not exist so all branch
specific checkouts were falling back to master which caused job
failures.

Instead just check if the base branch name exists in the repo as that
existing in a zuulv3 repo is sufficient to know we are up to date.

Depends-On: I110879bac5d8a671d73fab906ffbcf8156b1135d
Change-Id: Ifcae3ff217259926a8950f6da83212ee311002b2