Use = for string comparisons
Change I384e4c376076051e7a4a8b2d704107cbe3ac08ad introduced
comparisons of $DEVSTACK_PROJECT_FROM_GIT against a couple of
hardcoded string values in a bash [[ ]] expression, but incorrectly
used the -eq integer comparison operator. This seems to work under
normal circumstances, but actually breaks if the string value of the
variable contains a "." since the equivalence test tries to fail
early if it detects what it thinks is a float value (because it's
intended only for integers). Instead of -eq use = for string
comparisons here, because that's what the POSIX test(1) utility
mandates.

Change-Id: If33d7aea9142e650b57c0f58c14e2a4434e30bf0