Handle stestr execution for stackviz
This commit makes to handle a stestr execution for stackviz process.
Tempest is now in the process of switching the tempest run command to
use stestr internally instead of testrepository. This means that results
will be stored in .stestr instead of .testrepository. So, this commit
makes to handle a stestr or testrepository execution with checking which
directory exists before attempting to process the subunit results.

Change-Id: Ib19d3779deb89142e95762794dd1ce3fe3236b55