Check for virtualenv binary
The current check looks for a virtualenv package; but this isn't
portable to other platforms.  In the gate we also have installed from
upstream (to ensure latest) and the package might be on hold, etc
which might mess up the detection.

I think the KISS thing to do is to check for virtualenv and fail with
a useful message.

Change-Id: Icf15ddd18ca52885f5c4a89a7d83622a70c811d5