Remove setting of ETCD_DOWNLOAD_URL
Adding ETC binaries was a stop gap measure until packages were
available.  With the impending release of bionic that'll be soon :)

Also using the mirror on tarballs.o.o was unreliable so these binaries
are cached in the nodepool images as described in[1].  Having this
setting in d-g makes more work than is needed in the near term.
Removing it does introduce a new failure mode if an etcd version is
used in devstack but not yet cacched in the image.

[1] http://lists.openstack.org/pipermail/openstack-infra/2018-March/005871.html

Change-Id: I2a6649ae3094ca65aa31798ee9bddf2fcdd49260