Dynamically determine overlay MTU
Some clouds we run on no longer have MTUs of 1500 on their interfaces
and are smaller. We dynamically determine the MTU we should use on the
overlay networks. List all "real" interfaces with a permament mac addr
then use the smallest MTU less 50 bytes for vxlan overhead.

We only consider "real" interfaces so that we can run this multiple
times for separate overlays without perpetually decreasing the MTU due
to finding the lower MTU we just configured.

Change-Id: If163bf211570e690e255cc3ef6888d9e0cba19b5