Add gate job for stable/stein
Now we have stable/stein branch ready for devstack
and so does for all service projects.

This commit adds the stable/stein job running for
d-g gate.

NOTE: Adding legacy job vesion of stable/stein because
only legacy jobs use devstack-gate which is what we need
to test.

This commit removes the all stable zuulv3 native jobs
which does not use devstack-gate and so does not required
to be tested here.

Change-Id: I546038fa4164a3085b33fa0ef30a51ec237bdc3f