Replace git.openstack.org URLs with opendev.org URLs
Thorough replacement of git.openstack.org URLs with their opendev.org
counterparts.

Port of the present changes are needed to fix some errors in legacy jobs when
opendev.org is used as git base URL, as devstack-gate tries to clone
openstack-dev/devstack and it fails (the redirect works only when
git.openstack.org is used).

Change-Id: I81f6e00aa10ef7f93d0dead2f76369035a18a81b