Note status of CINDER_PERIODIC_INTERVAL setting
Change I1ef1fe564123216b19582262726cdb1078b7650e in DevStack removes
the ability to set the Cinder periodic_interval config option via the
CINDER_PERIODIC_INTERVAL shell var.

The CINDER_PERIODIC_INTERVAL was introduced in DevStack in Havana as
a workaround for Bug #1180976.  That bug, however, was fixed in such
a way that it does not require adjusting the periodic_interval in
Cinder, and the workaround was finally removed in Train.

Because devstack-gate is basically frozen as a legacy pre-zuul-v3
supporting infrastructure for DevStack jobs, this patch does not remove
setting the CINDER_PERIODIC_INTERVAL.  It simply adds a note explaining
that it's a no-op when using a Train or later DevStack.

Depends-on: https://review.opendev.org/#/c/652711
Change-Id: I7e2cc815ad14036dbae94e73d118b2845ddf5c64
Partial-bug: #1824837