Bump ANSIBLE_VERSION default to 2.7.14
devstack-gate is still used for grenade jobs and restricts
ara<1.0.0 and defaults ANSIBLE_VERSION=2.5.7.

ARA 0.16.6 dropped support for ansible < 2.6 [1]. As such,
grenade jobs are failing because ara 0.16.6 cannot be installed.

This bumps the ANSIBLE_VERSION default to 2.7.14 which is current
as of Ubuntu Xenial 16.04 which should be sufficient for the
branches still supported in upstream OpenStack CI (mainly Ocata+).

[1] https://github.com/ansible-community/ara/commit/872ba818fbc5267b1f769d5485fd3ab318235aea

Story: 2007028
Task: 37825

Change-Id: If7874f6e74606fe8a79fbff48abf764cc41c4a3b