Modernize default services in the test matrix
Ceilometer is not installed by default in legacy jobs using
devstack-gate so we shouldn't be enabling it by default. Anything
that cares about testing with ceilometer should be using a zuulv3
based job with the ceilometer devstack plugin and explicitly
enabling the service. So ceilometer is removed from the default
list of enabled services.

This also removes some now-dead nova services (n-obj, ha).

Change-Id: If8237c94be2b1db210f83bc9871792af9c5c868f