Remove 'cinder' service from feature test matrix
The cinder config is including the "cinder" devstack
service which would mean that anything removed explicitly,
like c-bak, is going to be trumped by the fact that 'cinder'
is in the enabled services list in a job.

This removes 'cinder' the service from cinder the config so
that cinder services have to be explicitly enabled/disabled
in the feature test matrix.

Change-Id: I857fbeba2af4a0f038fd2c2a19bb29c2767227a2