Drop peakmem_tracker service
When devstack reverted the pip cap in [0], it was necessary to disable
the peakmem_tracker service as there was no easy way to make it work in
that setup. So we need to delete it from the test-matrix here, too.

We could decide to add it only for older stable/branches, but as ds-gate
is to be obsoleted anyway, I'm not convinced that that is worth the
effort.

[0] https://review.opendev.org/561597

Change-Id: I9dd57d3bdb56c64377efbdd55429bd075223c746