Use stable constraints in Tempest tox till stable/rocky
Stable branches till stable/rocky is using python
version <py3.6. Tempest test those branch in venv
but Tempest tox use the master upper-constraint[1]
which block installation due to dependencies
require >=py3.6. For exmaple, oslo.concurrency 4.0.0
is not compatible for <py3.6.

As we pin Tempest for EM stable brach, we should be
able to use stable constraint for Tempest installation
as well as while running during run-tempest playbook.

tox.ini is hard coded to use master constraint[1] which force
legacy ocata job to use the master constraints and fail
- https://review.opendev.org/#/c/706161/

Devstack can set stable u-c to use via env var but this is
not enough as Tempest recreate the tox env with master u-c
when run test via devstack-gate.

All other branches have been fixed by fixing the run-tempest playbooks
in same way.
- https://review.opendev.org/#/c/705870/

Because stable/ocata and grenade job use devstack-gate, we need to fix this.

This is required to fix the grenade jobs on each stable branch to
get the other fixes in.

[1] https://opendev.org/openstack/tempest/src/commit/bc9fe8eca801f54915ff3eafa418e6e18ac2df63/tox.ini#L14

Change-Id: Ibfc5c3d5e64e274fd674e6c42878d98bb194fa73