Convert testmatrix run to venv
As part of our efforts to ship bare nodes, we want to install as
little as possible in any sort of pre-run jobs.  Since all infra nodes
have python3 available, use venv to create this small environment for
the test matrix production, and avoid having to pull in any other
dependencies.

Change-Id: Ie8b09ab60841f9d1b15ae1a7afec51906cd89f95