Switch legacy-tempest-neutron-full-stable to focal
legacy-tempest-neutron-full-stable deploy's stable/wallaby
and is running on ubuntu-bionic(py3.6) and failing in tempest
setup as tempest master no longer works with py3.6.

Master is dropping py3.6 support, currently oslo.log has already
dropped the support and it's latest release is not installable
on py3.6.

Pin Markusafe, as since version 2.1.0 it dropped
module 'soft_unicode' which is required by Jinja2.

Also update ansible version to 2.9.6 as of focal
which should auto discover python as per distro.

So let's switch this job to ubuntu-focal which is py3.8.

Change-Id: I37887bf430b4bce2b9d2179b2d6dd6c050bb1768