Use stable constraints for Tempest venv for stable/ussuri|victoria
we do have legacy jobs in ussuri and might be in victoria also and
as we are pinning tempest in both stable branches they need to use the
respective stable branches constraints instead of master constraints.

- https://review.opendev.org/q/topic:ussuri-pin-tempest
- https://review.opendev.org/q/topic:victoria-pin-tempest

Hoping this is last time we are updating the tempest setting in d-g
as all jobs in victoria onwards should be migrated to zuulv3 native.

Change-Id: I07e1cb9b038e8d0c8ce83e8554188f9eb72cb339