Retire devstack gate
This "removes" devstack-gate content in order to retire it. Devstack
grew the ability to bootstrap its own CI environments using ansible and
no longer needs devstack-gate.

Part of the motivation for this chagne is that it will help us in our
quest to remove old Ubuntu Xenial test nodes from the CI system as
devstack-gate still relies on them.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/919625
Change-Id: Ife60f1dd6fae7577cee78054b69d8ab83df9a8ce